Types adjustment, extract captions to a function
